Enhance graph title styling
