+++ packages/nexusgraph-app/src/component/GraphTitle.tsx
@@ -1,0 +1,24 @@
+// Copyright 2023 Paion Data. All rights reserved.
+import { useEffect, useState } from "react";
+import { selectGraphData } from "../../../nexusgraph-redux";
+import { StyledInput } from "./styled";
+
+export default function GraphTitle(): JSX.Element {
+  const graphData = selectGraphData();
+
+  const [inputValue, setInputValue] = useState(graphData.name);
+
+  useEffect(() => {
+    setInputValue(graphData.name);
+  }, [graphData.id]);
+
+  return (
+    <StyledInput
+      id="titleInput"
+      value={inputValue}
+      onChange={(event) => {
+        setInputValue(event.target.value);
+      }}
+    />
+  );
+}
